backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langchain==0.3.7
langchain-community==0.3.7
langchain-core>=0.3.17,<0.4.0
chromadb==0.5.15
rank-bm25==0.2.2
sentence-transformers==3.3.0
pymupdf==1.24.11
python-docx==1.1.2
httpx==0.27.2
numpy==1.26.4
pydantic==2.9.2
